mcps/servers/gmail/pyproject.toml
[project]
name = "gmail-mcp-server"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastmcp==3.0.2",
    "google-api-python-client==2.190.0",
    "google-auth-httplib2==0.3.0",
    "google-auth-oauthlib==1.3.0",
]
